Fix fieldsetCards initialization - remove incorrect hasHeadFields check
The hasHeadFields check was preventing cards mode from working because
it was looking for data-head-* attributes in the wrong place. In cards
mode, headTitle and headPreview are configured in the Resource but the
attributes are on individual card headers, not on a container-level element.

Removed the check - cards view is identified by .fieldset-cards-view class,
and that's sufficient. Also removed debug console.log statements.

This fixes the issue where clicking on card items did nothing.